upload the ogm study
The study for regional purchasing power conducted by OGM (2009) is often cited in literature about regional purchasing power. Furthermore, I partly build on this methodology and refer to it's results, since they are for the same regions. Unfortunately, the study is no longer publicly available (researchers who cite it got it from the old OGM homepage which was revised in 2020) so I upload it again to better reconstruct my results.